Remove some bad code. Provide menu bar on login screen.
