+++ client/src/main/java/ru/ryabtsev/cloud/client/LoginController.java
@@ -4,6 +4,7 @@ import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.PasswordField;
 import javafx.scene.control.TextField;
+import javafx.scene.layout.BorderPane;
 import javafx.scene.layout.VBox;
 import ru.ryabtsev.cloud.client.service.AuthencticationService;
 import ru.ryabtsev.cloud.client.service.NetworkAuthenticationService;
@@ -19,7 +20,7 @@ public class LoginController {
     PasswordField password;
 
     @FXML
-    VBox parentWindow;
+    BorderPane parentWindow;
 
     private AuthencticationService authencticationService = new NetworkAuthenticationService(
             ClientApplication.networkService()
